Keep command-specific printers with their command implementations.
