+++ internal/command/sandbox/get.go
@@ -48,7 +48,7 @@ func get(cfg *config.SandboxGet, out io.Writer, name string) error {
 
 	switch cfg.OutputFormat {
 	case config.OutputFormatDefault:
-		return print.SandboxDetails(cfg, out, sb)
+		return printSandboxDetails(cfg.Sandbox, out, sb)
 	case config.OutputFormatJSON:
 		return print.RawJSON(out, sb)
 	case config.OutputFormatYAML:
